processing/cloud_function/D001/structure/structure-job/requirements.txt
pydantic==2.9.2
azure-ai-documentintelligence==1.0.0
pandas==2.2.2
python-magic==0.4.27
boto3==1.35.74
langchain-aws==0.2.10
requests==2.32.3
langchain-community==0.3.14
aiofiles==24.1.0
motor==3.6.0
openpyxl